Lot of small failures fixed
* Fixed include statement error when using brakets where quotes are mandatary.

* Improvement of the executor adding the parameters parameter. With parameters parameter you can add lots of parameters without the need of implements new execution methods.

+++ v3/agent/LowLevel.cpp
@@ -261,9 +261,9 @@ int Device::GetMinorVersion()
 /*!
 	@brief Gets the total global memory available on the device, in bytes.
  */
-unsigned int Device::GetTotalMem()
+size_t Device::GetTotalMem()
 {
-	unsigned int mem;
+	size_t mem;
 	CUresult err;
 	if(CUDA_SUCCESS != (err=cuDeviceTotalMem(&mem, m_device)))
 		ThrowCudaLLError("Failed to get CUDA device memory size", err);
